Delete core/tee/tee_fs.c
tee_fs.c contains trivial wrappers, that will eventually go. Move them
temporarily into tee_rpmb_fs.c and tee_fs_common.c (which will later be
renamed tee_ree_fs.c).

Tested-by: Pascal Brand <pascal.brand@linaro.org> (STM)
