+++ core/include/tee/tee_fs.h
@@ -34,31 +34,6 @@
 
 #define TEE_FS_NAME_MAX 350
 
-/*
- * We split a TEE file into multiple blocks and store them
- * on REE filesystem. A TEE file is represented by a REE file
- * called meta and a number of REE files called blocks. Meta
- * file is used for storing file information, e.g. file size
- * and backup version of each block.
- *
- * REE files naming rule is as follows:
- *
- *   <tee_file_name>/meta.<backup_version>
- *   <tee_file_name>/block0.<backup_version>
- *   ...
- *   <tee_file_name>/block15.<backup_version>
- *
- * Backup_version is used to support atomic update operation.
- * Original file will not be updated, instead we create a new
- * version of the same file and update the new file instead.
- *
- * The backup_version of each block file is stored in meta
- * file (see @backup_version_table in tee_fs_private.h), the
- * meta file itself also has backup_version, the update is
- * successful after new version of meta has been written.
- */
-#define REE_FS_NAME_MAX (TEE_FS_NAME_MAX + 20)
-
 typedef int64_t tee_fs_off_t;
 typedef uint32_t tee_fs_mode_t;
 typedef struct tee_fs_dir tee_fs_dir;
